+++ cancer_api/files.py
@@ -5,13 +5,15 @@ This submodules contains all classes representing file
 types/formats, which in turn employ the parsers submodule.
 """
 
-import gzip
 import os.path
 import gc
+import logging
 import parsers
 import mutations
 import misc
-from exceptions import CancerApiException, IllegalVariableDefinition
+from base import CancerApiObject
+from exceptions import CancerApiException
+from utils import open_file
 
 
 class BaseFile(object):
@@ -19,80 +21,83 @@ class BaseFile(object):
 
     DEFAULT_HEADER = ""
     HEADER_PREFIX = "#"
-    FILE_EXTENSIONS = []
+    FILE_EXTENSIONS = ["txt"]
 
     def __init__(self, *args, **kwargs):
-        """Provide more informative error message.
-        Redirect users to open and convert methods.
-        """
-        raise CancerApiException("Cannot instantiate file directly. "
-                                 "Please use `open` and `convert` methods instead.")
+        """Can't initialize directly."""
+        raise CancerApiException("Please use `open`, `convert` or `new` methods instead.")
 
     @classmethod
-    def open(cls, filepath, parser_cls=None):
-        """Instantiate a BaseFile object from an
-        existing file on disk.
+    def _init(cls, filepath=None, parser_cls=None, other_file=None, is_new=False, buffersize=None):
+        """Initialize BaseFile. Any instantiation of BaseFile should
+        go through this method in an attempt to standardize attributes.
+        Meant to be used internally only.
         """
         obj = cls.__new__(cls)
         obj.filepath = filepath
-        if not parser_cls:
-            parser_cls = cls.DEFAULT_PARSER_CLS
-        obj.parser = parser_cls(obj)
-        obj._source = obj
-        obj.is_new_file = False
+        obj.parser_cls = parser_cls
+        obj.parser = parser_cls(obj) if parser_cls else cls.DEFAULT_PARSER_CLS(obj)
+        obj.source = obj if not other_file else other_file.source
+        obj.is_new = is_new
+        obj.storelist = []
+        obj.buffersize = buffersize
         return obj
 
     @classmethod
-    def convert(cls, other_file):
+    def open(cls, filepath, parser_cls=None, buffersize=None):
+        """Instantiate a BaseFile object from an
+        existing file on disk.
+        """
+        obj = cls._init(filepath=filepath, parser_cls=parser_cls, other_file=None, is_new=False,
+                        buffersize=buffersize)
+        return obj
+
+    @classmethod
+    def convert(cls, filepath, other_file, buffersize=None):
         """Instantiate a BaseFile object from another
         BaseFile object.
         """
         if not isinstance(other_file, BaseFile):
-            raise CancerApiException("Must pass cancer_api file object.")
-        obj = cls.__new__(cls)
-        obj._source = other_file._source
-        obj.is_new_file = True
+            raise CancerApiException("Must pass cancer_api file object as `other_file`.")
+        obj = cls._init(filepath=filepath, parser_cls=None, other_file=other_file, is_new=True,
+                        buffersize=buffersize)
+        obj.write()
         return obj
 
     @classmethod
-    def new(cls, filepath=None):
+    def new(cls, filepath, buffersize=None):
         """Instantiate a BaseFile object from scratch.
         Useful for adding objects and write them out to disk.
         """
-        obj = cls.__new__(cls)
-        obj.filepath = filepath
-        obj._source = obj
-        obj.is_new_file = True
+        obj = cls._init(filepath=filepath, parser_cls=None, other_file=None, is_new=True,
+                        buffersize=buffersize)
         return obj
 
-    @property
-    def header(self):
+    def get_header(self):
         """Return header if already stored in instance.
         Otherwise, parse file on disk if filepath is specified.
         If not, return default header for current file type.
         """
-        if getattr(self, "_header", None):
-            header = self._header
-        elif getattr(self, "_filepath", None):
-            with self._open(self.filepath) as infile:
+        if self.is_new:
+            # It's a file create with the `new` or `convert` methods
+            header = self.DEFAULT_HEADER
+        else:
+            # It's a file created with the `open` method
+            with self._open() as infile:
                 header = ""
                 for line in infile:
                     if self.is_header_line(line):
                         header += line
-        else:
-            header = self.DEFAULT_HEADER
+                    else:
+                        break
         return header
-
-    @header.setter
-    def header(self, value):
-        self._header = value
 
     @property
     def col_names(self):
         if getattr(self, "_col_names", None):
             col_names = self._col_names
         else:
-            header = self.header
+            header = self.get_header()
             # Assume that the column names are the
             # last line in the header
             last_header_line = header.rstrip("\n").split("\n")[-1]
@@ -104,58 +109,22 @@ class BaseFile(object):
     def col_names(self, value):
         self._col_names = list(value)
 
-    @property
-    def filename(self):
-        if self._source == self:
-            full_filename = os.path.basename(self.filepath)
-            for ext in self.FILE_EXTENSIONS:
-                if full_filename.lower().endswith("." + ext.lower()):
-                    ext_len = len("." + ext)
-                    return full_filename[:-ext_len]
-            return full_filename
-        else:
-            return None
-
-    @filename.setter
-    def filename(self, value):
-        raise IllegalVariableDefinition("Cannot directly edit filename.")
+    def split_filename(self):
+        """Returns filename (root, ext) tuple."""
+        filename = os.path.basename(self.filepath)
+        # Make sure that long extensions go first (to match the longest available extension)
+        for ext in ("." + x.lower() for x in sorted(self.FILE_EXTENSIONS, key=len, reverse=True)):
+            if filename.lower().endswith(ext):
+                return (filename[:-len(ext)], ext)
+        # If none of the class' file extensions match, just use os.path.splitext
+        return os.path.splitext(filename)
 
     @classmethod
-    def get_extension(cls):
+    def get_cls_extension(cls):
         """Return first extension from cls.FILE_EXTENSIONS.
         Otherwise, returns an error if not available.
         """
-        if not getattr(cls, "FILE_EXTENSIONS", None):
-            raise CancerApiException("{} doesn't have file extensions specified "
-                                     "yet.".format(cls.__name__))
         return cls.FILE_EXTENSIONS[0]
-
-    @property
-    def filepath(self):
-        return self._source._filepath
-
-    @filepath.setter
-    def filepath(self, value):
-        self._filepath = value
-
-    @property
-    def parser(self):
-        return self._source._parser
-
-    @parser.setter
-    def parser(self, value):
-        self._parser = value
-
-    @property
-    def storelist(self):
-        if not getattr(self, "_storelist", None):
-            self._storelist = []
-        return self._storelist
-
-    @storelist.setter
-    def storelist(self, value):
-        raise IllegalVariableDefinition("Modify storelist using methods such as "
-                                        "`add_obj`, `rm_obj`, `clear_storelist`")
 
     def add_obj(self, obj):
         """Add object to storelist. Useful to bind objects
@@ -163,17 +132,16 @@ class BaseFile(object):
         Returns whether the object was added
         (always True for now).
         """
-        # Ensure that storelist is defined
-        self.storelist
-        self._storelist.append(obj)
+        if not isinstance(obj, CancerApiObject):
+            raise CancerApiException("`add_obj` only supports cancer_api objects")
+        self.storelist.append(obj)
+        if self.buffersize and len(self.storelist) >= self.buffersize:
+            self.write()
         return True
-
-    def rm_obj(self, obj):
-        raise NotImplementedError("`BaseFile.rm_obj` is not implemented yet.")
 
     def clear_storelist(self):
         """Empty storelist"""
-        self._storelist = []
+        self.storelist = []
         # Force garbage collection
         gc.collect()
 
@@ -181,32 +149,13 @@ class BaseFile(object):
     def is_header_line(cls, line):
         """Return whether or not a line is a header line
         according to the current file type.
-        Defaults to lines starting with '#'.
+        Defaults to lines starting with '#'
+        (see BaseFile.HEADER_PREFIX).
         """
         is_header_line = False
-        # If HEADER_PREFIX is None, always return false
-        if cls.HEADER_PREFIX is not None and line.startswith(cls.HEADER_PREFIX):
+        if cls.HEADER_PREFIX and line.startswith(cls.HEADER_PREFIX):
             is_header_line = True
         return is_header_line
-
-    def _open(self, filepath, mode="r", *args, **kwargs):
-        """Wrapper for file open() function.
-        Purpose: to catch gzipped files and handle them
-        accordingly by using the gzip module.
-        """
-        opened_file = None
-        if filepath.endswith(".gz"):
-            # Ensure that "b" is in the mode
-            if "b" not in mode:
-                mode += "b"
-            # Override compression level default (9 -> 6)
-            # if not specified
-            if "compresslevel" not in kwargs:
-                kwargs["compresslevel"] = 6
-            opened_file = gzip.open(filepath, mode, *args, **kwargs)
-        else:
-            opened_file = open(filepath, mode, *args, **kwargs)
-        return opened_file
 
     @classmethod
     def obj_to_str(cls, obj):
@@ -225,64 +174,72 @@ class BaseFile(object):
         # return line
         raise NotImplementedError
 
-    def write(self, outfilepath=None):
-        """Write BaseFile instance out to disk.
-        If not outfilepath specified, appends objects
-        stored in self.storelist to current filepath.
-        Will iterate over source's file and any objects
-        stored in self.storelist.
-        If instance is a converted file, update
-        filepath and parser attributes accordingly.
+    def _open(self):
+        """Use the open_file function on self.source.filepath in 'r' mode"""
+        return open_file(self.source.filepath)
+
+    def write(self, outfilepath=None, mode="w"):
+        """Write objects in file to disk.
+        Either you can write to a new file (if outfilepath is given),
+        or you can append what's in storelist to the current filepath.
         """
-        # If outfilepath is specified, write every object in
-        # self.source and self.storelist to disk.
-        # Clear self.storelist afterwards.
+        # If outfilepath is specified, iterate over every object in self
+        # (which might come from something else) and every object in
+        # self.storelist and write them out to disk
         if outfilepath:
-            with self._open(outfilepath, "w") as outfile:
-                outfile.write(self.header)
-                for obj in self._source:
+            with open_file(outfilepath, mode) as outfile:
+                logging.info("Writing to disk...")
+                outfile.write(self.get_header())
+                for obj in self.source:
                     line = self.obj_to_str(obj)
                     outfile.write(line)
-                for obj in self._source.storelist:
-                    line = self.obj_to_str(obj)
-                    outfile.write(line)
-            if self._source is not self:
-                self._source = self
-                self.filepath = outfilepath
-                self.is_new_file = False
-                self.parser = self.DEFAULT_PARSER_CLS
-            self.clear_storelist()
-        # If outfilepath is not specified, it is assumed that the
-        # file instance already has self.filepath and the purpose
-        # is to append any objects stored in self.storelist to the
-        # file on disk.
-        # Clear self.storelist afterwards.
-        else:
-            if self._source is not self or not self.filepath:
-                raise CancerApiException("This file has never been written out to disk. "
-                                         "Please specify an outfilepath at first.")
-            # Check if this is a new file and if self.filepath exists
-            if self.is_new_file and os.path.exists(self.filepath):
-                raise CancerApiException("The specified filepath already exists. If you wish "
-                                         "to overwrite the file, delete it or explicitly "
-                                         "specify an outfilepath for the `write` method.")
-            with self._open(self.filepath, "a+") as outfile:
                 for obj in self.storelist:
                     line = self.obj_to_str(obj)
                     outfile.write(line)
+            # Clear storelist now that they've been written to disk
             self.clear_storelist()
+            # Update file attributes (in case of new or converted file)
+            self.source = self
+            self.filepath = outfilepath
+            self.is_new = False
+
+        # If outfilepath is not specified, simply iterate over every
+        # object in self.storelist and append them to the file on disk.
+        else:
+            # If the file is new and the path already exist, do not append
+            if self.is_new and os.path.exists(self.filepath):
+                raise CancerApiException("Output file already exists: {}".format(self.filepath))
+            with open_file(self.filepath, "a+") as outfile:
+                logging.info("Writing to disk...")
+                # If the file is new, start with header
+                if self.is_new:
+                    outfile.write(self.get_header())
+                # If file is new and source is not self (i.e., converted file),
+                # iterate over source
+                if self.is_new and self.source is not self:
+                    for obj in self.source:
+                        line = self.obj_to_str(obj)
+                        outfile.write(line)
+                # Proceed with iterating over storelist
+                for obj in self.storelist:
+                    line = self.obj_to_str(obj)
+                    outfile.write(line)
+            # Clear storelist now that they've been written to disk
+            self.clear_storelist()
+            # Update file attributes (in case of new or converted file)
+            self.source = self
+            self.is_new = False
 
     def __iter__(self):
         """Return instances of the objects
         associated with the current file type.
         """
-        # Iterate over every non-header line
-        source = self._source
-        with self._open(source.filepath) as infile:
+        # Iterate over every non-header line in self.source
+        with self._open() as infile:
             for line in infile:
-                if source.is_header_line(line):
+                if self.source.is_header_line(line):
                     continue
-                obj = source.parser.parse(line)
+                obj = self.source.parser.parse(line)
                 if obj:
                     yield obj
 
@@ -298,7 +255,7 @@ class BedpeFile(BaseFile):
     """Class for representing BEDPE files."""
 
     FILE_EXTENSIONS = ["bedpe"]
-    DEFAULT_HEADER = "chrom1\tstart1\tend1\tchrom2\tstart2\tend2\tname\tscore\tstrand1\tstrand2\n"
+    DEFAULT_HEADER = "#chrom1\tstart1\tend1\tchrom2\tstart2\tend2\tname\tscore\tstrand1\tstrand2\n"
     DEFAULT_PARSER_CLS = parsers.BaseParser  # A BEDPE parser need to be implemented
 
     @classmethod
@@ -345,12 +302,11 @@ class FastqFile(BaseFile):
         quartets (groups of four lines).
         """
         # Iterate over quartets (non-header lines)
-        source = self._source
-        with self._open(source.filepath) as infile:
+        with self._open() as infile:
             current_quartet = ""
             for line in infile:
                 # Skip header lines
-                if source.is_header_line(line):
+                if self.source.is_header_line(line):
                     continue
                 # Add line to quartet
                 current_quartet += line
@@ -360,7 +316,7 @@ class FastqFile(BaseFile):
                     continue
                 # Once the quartet has four lines, it will continue on
                 # to being parsed
-                obj = source.parser.parse(current_quartet)
+                obj = self.source.parser.parse(current_quartet)
                 if obj:
                     yield obj
                 current_quartet = ""